fix: 20 Kotlin compile errors blocking CI build
Files fixed:
- Theme.kt: palette vals declared before DarkColorScheme (forward ref fix)
- WmsTileInterceptor.kt: import okhttp3.Response (unresolved ref)
- MapStyleBuilder.kt: TileSet uses single String, not list; remove unused import
- LocationService.kt: .setMaxUpdates takes Int, not Long
- ToiletSearchApi.kt: SEARCH_RADIUS.toDouble() for Double param
- Models.kt: replace .pow(2) with .let { it * it } (no pow on Double)
- MapScreen.kt: cameraPosition → nativeMap.cameraPosition; add true return
  on tap/long-press listeners (expect Boolean, not Job/Unit)
- RoutePlannerModal.kt: typography.caption → labelSmall (M3 rename)
- SettingsPanel.kt: padding(start=..., vertical=...) → top=,bottom=
  (no vertical param in that overload)

+++ app/src/main/java/com/turbolego/rullut/api/LocationService.kt
@@ -124,7 +124,7 @@ object LocationService {
             Priority.PRIORITY_HIGH_ACCURACY,
             UPDATE_INTERVAL_MS,
         ).setMinUpdateIntervalMillis(FASTEST_INTERVAL_MS)
-            .setMaxUpdates(Int.MAX_VALUE.toLong())
+            .setMaxUpdates(Int.MAX_VALUE)
             .build()
 
         try {
